refactor: split svg parser modules
