+++ src/parser/svg/model.rs
@@ -1,0 +1,84 @@
+use super::style::SvgStyle;
+
+/// A parsed SVG tree ready for rendering.
+#[derive(Debug, Clone)]
+pub struct SvgTree {
+    pub width: f32,
+    pub height: f32,
+    pub width_attr: Option<String>,
+    pub height_attr: Option<String>,
+    pub view_box: Option<ViewBox>,
+    pub children: Vec<SvgNode>,
+}
+
+#[derive(Debug, Clone)]
+pub struct ViewBox {
+    pub min_x: f32,
+    pub min_y: f32,
+    pub width: f32,
+    pub height: f32,
+}
+
+#[derive(Debug, Clone)]
+pub enum SvgNode {
+    Group {
+        transform: Option<SvgTransform>,
+        children: Vec<SvgNode>,
+        style: SvgStyle,
+    },
+    Rect {
+        x: f32,
+        y: f32,
+        width: f32,
+        height: f32,
+        rx: f32,
+        ry: f32,
+        style: SvgStyle,
+    },
+    Circle {
+        cx: f32,
+        cy: f32,
+        r: f32,
+        style: SvgStyle,
+    },
+    Ellipse {
+        cx: f32,
+        cy: f32,
+        rx: f32,
+        ry: f32,
+        style: SvgStyle,
+    },
+    Line {
+        x1: f32,
+        y1: f32,
+        x2: f32,
+        y2: f32,
+        style: SvgStyle,
+    },
+    Polyline {
+        points: Vec<(f32, f32)>,
+        style: SvgStyle,
+    },
+    Polygon {
+        points: Vec<(f32, f32)>,
+        style: SvgStyle,
+    },
+    Path {
+        commands: Vec<PathCommand>,
+        style: SvgStyle,
+    },
+}
+
+#[derive(Debug, Clone)]
+pub enum SvgTransform {
+    Matrix(f32, f32, f32, f32, f32, f32),
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum PathCommand {
+    MoveTo(f32, f32),
+    LineTo(f32, f32),
+    CubicTo(f32, f32, f32, f32, f32, f32),
+    QuadTo(f32, f32, f32, f32),
+    ClosePath,
+}
